[MOD] move merge test from redis_api_test.go to redis_api_merge_test.go
