+++ server/redis_api_test.go
@@ -20,7 +20,6 @@ var testOnce sync.Once
 var kvs *Server
 var redisport int
 var OK = "OK"
-var kvsmutil *Server
 
 func startTestServer(t *testing.T) (*Server, int, string) {
 	tmpDir, err := ioutil.TempDir("", fmt.Sprintf("rocksdb-test-%d", time.Now().UnixNano()))
@@ -1869,254 +1868,4 @@ func TestZSetScan(t *testing.T) {
 	} else {
 		checkScanValues(t, ay[1], "a", 1, "b", 2)
 	}
-}
-
-func startMergeTestServer(t *testing.T) (*Server, int, string) {
-	tmpDir, err := ioutil.TempDir("", fmt.Sprintf("rocksdb-test-%d", time.Now().UnixNano()))
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Logf("dir:%v\n", tmpDir)
-	ioutil.WriteFile(
-		path.Join(tmpDir, "myid"),
-		[]byte(strconv.FormatInt(int64(1), 10)),
-		common.FILE_PERM)
-	raftAddr := "http://127.0.0.1:12345"
-	redisport := 22345
-	var replica node.ReplicaInfo
-	replica.NodeID = 1
-	replica.ReplicaID = 1
-	replica.RaftAddr = raftAddr
-	kvOpts := ServerConfig{
-		ClusterID:     "test",
-		DataDir:       tmpDir,
-		RedisAPIPort:  redisport,
-		LocalRaftAddr: raftAddr,
-		BroadcastAddr: "127.0.0.1",
-		TickMs:        100,
-		ElectionTick:  5,
-	}
-	nsConf := node.NewNSConfig()
-	nsConf.Name = "default-0"
-	nsConf.BaseName = "default"
-	nsConf.EngType = rockredis.EngType
-	nsConf.PartitionNum = 3
-	nsConf.Replicator = 1
-	nsConf.RaftGroupConf.GroupID = 1000
-	nsConf.RaftGroupConf.SeedNodes = append(nsConf.RaftGroupConf.SeedNodes, replica)
-	kv := NewServer(kvOpts)
-	_, err = kv.InitKVNamespace(1, nsConf, false)
-	if err != nil {
-		t.Fatalf("failed to init namespace: %v", err)
-	}
-	nsConf1 := node.NewNSConfig()
-	nsConf1.Name = "default-1"
-	nsConf1.BaseName = "default"
-	nsConf1.EngType = rockredis.EngType
-	nsConf1.PartitionNum = 3
-	nsConf1.Replicator = 1
-	nsConf1.RaftGroupConf.GroupID = 1000
-	nsConf1.RaftGroupConf.SeedNodes = append(nsConf.RaftGroupConf.SeedNodes, replica)
-	_, err = kv.InitKVNamespace(1, nsConf1, false)
-	if err != nil {
-		t.Fatalf("failed to init namespace: %v", err)
-	}
-
-	nsConf2 := node.NewNSConfig()
-	nsConf2.Name = "default-2"
-	nsConf2.BaseName = "default"
-	nsConf2.EngType = rockredis.EngType
-	nsConf2.PartitionNum = 3
-	nsConf2.Replicator = 1
-	nsConf2.RaftGroupConf.GroupID = 1000
-	nsConf2.RaftGroupConf.SeedNodes = append(nsConf.RaftGroupConf.SeedNodes, replica)
-	_, err = kv.InitKVNamespace(1, nsConf2, false)
-	if err != nil {
-		t.Fatalf("failed to init namespace: %v", err)
-	}
-
-	kv.Start()
-	time.Sleep(time.Second)
-	return kv, redisport, tmpDir
-}
-
-func getMergeTestConn(t *testing.T) *goredis.PoolConn {
-	testOnce.Do(func() {
-		kvsmutil, redisport, _ = startMergeTestServer(t)
-	},
-	)
-	c := goredis.NewClient("127.0.0.1:"+strconv.Itoa(redisport), "")
-	c.SetMaxIdleConns(4)
-	conn, err := c.Get()
-	if err != nil {
-		t.Fatal(err)
-	}
-	return conn
-}
-
-func checkMergeScanValues(t *testing.T, ay interface{}, values ...interface{}) {
-	a, err := goredis.Strings(ay, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(a) != len(values) {
-		t.Fatal(fmt.Sprintf("len %d != %d", len(a), len(values)))
-	}
-	equalCount := 0
-
-	for _, v := range a {
-		for j, _ := range values {
-			if string(v) == fmt.Sprintf("%v", values[j]) {
-				equalCount++
-			}
-		}
-	}
-
-	if equalCount != len(a) {
-		t.Fatal("not all equal")
-	}
-}
-
-func checkMergeAdvanceScan(t *testing.T, c *goredis.PoolConn, tp string) {
-	if ay, err := goredis.Values(c.Do("ADVSCAN", "default:testscan:", tp, "count", 5)); err != nil {
-		t.Fatal(err)
-	} else if len(ay) != 2 {
-		t.Fatal(len(ay))
-	} else if n := ay[0].([]byte); string(n) != "MDpkR1Z6ZEhOallXNDZNUT09OzE6ZEdWemRITmpZVzQ2TUE9PTsyOmRHVnpkSE5qWVc0Nk1URT07" &&
-		string(n) != "MDpkR1Z6ZEhOallXNDZNUT09OzI6ZEdWemRITmpZVzQ2TVRFPTsxOmRHVnpkSE5qWVc0Nk1BPT07" &&
-		string(n) != "MTpkR1Z6ZEhOallXNDZNQT09OzA6ZEdWemRITmpZVzQ2TVE9PTsyOmRHVnpkSE5qWVc0Nk1URT07" &&
-		string(n) != "MTpkR1Z6ZEhOallXNDZNQT09OzI6ZEdWemRITmpZVzQ2TVRFPTswOmRHVnpkSE5qWVc0Nk1RPT07" &&
-		string(n) != "MjpkR1Z6ZEhOallXNDZNVEU9OzA6ZEdWemRITmpZVzQ2TVE9PTsxOmRHVnpkSE5qWVc0Nk1BPT07" &&
-		string(n) != "MjpkR1Z6ZEhOallXNDZNVEU9OzE6ZEdWemRITmpZVzQ2TUE9PTswOmRHVnpkSE5qWVc0Nk1RPT07" {
-		t.Fatal(string(n))
-	} else {
-		checkMergeScanValues(t, ay[1], "testscan:0", "testscan:1", "testscan:11")
-	}
-
-	if ay, err := goredis.Values(c.Do("ADVSCAN", "default:testscan:MDpkR1Z6ZEhOallXNDZNUT09OzE6ZEdWemRITmpZVzQ2TUE9PTsyOmRHVnpkSE5qWVc0Nk1URT07", tp, "count", 6)); err != nil {
-		t.Fatal(err)
-	} else if len(ay) != 2 {
-		t.Fatal(len(ay))
-	} else if n := ay[0].([]byte); string(n) != "MDpkR1Z6ZEhOallXNDZNVFE9OzE6ZEdWemRITmpZVzQ2TVRVPTsyOmRHVnpkSE5qWVc0Nk1UZz07" &&
-		string(n) != "MDpkR1Z6ZEhOallXNDZNVFE9OzI6ZEdWemRITmpZVzQ2TVRnPTsxOmRHVnpkSE5qWVc0Nk1UVT07" &&
-		string(n) != "MTpkR1Z6ZEhOallXNDZNVFU9OzA6ZEdWemRITmpZVzQ2TVRRPTsyOmRHVnpkSE5qWVc0Nk1UZz07" &&
-		string(n) != "MTpkR1Z6ZEhOallXNDZNVFU9OzI6ZEdWemRITmpZVzQ2TVRnPTswOmRHVnpkSE5qWVc0Nk1UUT07" &&
-		string(n) != "MjpkR1Z6ZEhOallXNDZNVGc9OzA6ZEdWemRITmpZVzQ2TVRRPTsxOmRHVnpkSE5qWVc0Nk1UVT07" &&
-		string(n) != "MjpkR1Z6ZEhOallXNDZNVGc9OzE6ZEdWemRITmpZVzQ2TVRVPTswOmRHVnpkSE5qWVc0Nk1UUT07" {
-		t.Fatal(string(n))
-	} else {
-		checkMergeScanValues(t, ay[1], "testscan:12", "testscan:14", "testscan:10", "testscan:15", "testscan:13", "testscan:18")
-	}
-
-	if ay, err := goredis.Values(c.Do("ADVSCAN", "default:testscan:MDpkR1Z6ZEhOallXNDZNVFE9OzE6ZEdWemRITmpZVzQ2TVRVPTsyOmRHVnpkSE5qWVc0Nk1UZz07", tp, "count", 8)); err != nil {
-		t.Fatal(err)
-	} else if len(ay) != 2 {
-		t.Fatal(len(ay))
-	} else if n := ay[0].([]byte); string(n) != "MDpkR1Z6ZEhOallXNDZNVGs9OzI6ZEdWemRITmpZVzQ2TlE9PTs=" &&
-		string(n) != "MjpkR1Z6ZEhOallXNDZOUT09OzA6ZEdWemRITmpZVzQ2TVRrPTs=" {
-		t.Fatal(string(n))
-	} else {
-		if len(ay[1].([]interface{})) != 0 {
-			checkMergeScanValues(t, ay[1], "testscan:16", "testscan:19", "testscan:17", "testscan:2", "testscan:5")
-		}
-	}
-
-	if ay, err := goredis.Values(c.Do("ADVSCAN", "default:testscan:MDpkR1Z6ZEhOallXNDZNVGs9OzI6ZEdWemRITmpZVzQ2TlE9PTs=", tp, "count", 5)); err != nil {
-		t.Fatal(err)
-	} else if len(ay) != 2 {
-		t.Fatal(len(ay))
-	} else if n := ay[0].([]byte); string(n) != "MDpkR1Z6ZEhOallXNDZOQT09OzI6ZEdWemRITmpZVzQ2T1E9PTs=" &&
-		string(n) != "MjpkR1Z6ZEhOallXNDZPUT09OzA6ZEdWemRITmpZVzQ2TkE9PTs=" {
-		t.Fatal(string(n))
-	} else {
-		if len(ay[1].([]interface{})) != 0 {
-			checkMergeScanValues(t, ay[1], "testscan:9", "testscan:8", "testscan:3", "testscan:4")
-		}
-	}
-
-	if ay, err := goredis.Values(c.Do("ADVSCAN", "default:testscan:MDpkR1Z6ZEhOallXNDZOQT09OzI6ZEdWemRITmpZVzQ2T1E9PTs=", tp, "count", 8)); err != nil {
-		t.Fatal(err)
-	} else if len(ay) != 2 {
-		t.Fatal(len(ay))
-	} else if n := ay[0].([]byte); string(n) != "" &&
-		string(n) != "" {
-		t.Fatal(string(n))
-	} else {
-		if len(ay[1].([]interface{})) != 0 {
-			checkMergeScanValues(t, ay[1], "testscan:6", "testscan:7")
-		}
-	}
-
-	if _, err := goredis.Values(c.Do("ADVSCAN", "default::MDpkR1Z6ZEhOallXNDZOQT09OzI6ZEdWemRITmpZVzQ2T1E9PTs=", tp, "count", 8)); err == nil {
-		t.Fatal("want err, get nil ")
-	}
-
-	if _, err := goredis.Values(c.Do("ADVSCAN", "default:testscan1:MDpkR1Z6ZEhOallXNDZOQT09OzI6ZEdWemRITmpZVzQ2T1E9PTs=", tp, "count", 8)); err == nil {
-		t.Fatal("want err, get nil ")
-	}
-
-	if _, err := goredis.Values(c.Do("ADVSCAN", "default:testscan:dGVzdHNjYW46NA==", tp, "count", 0)); err == nil {
-		t.Fatal("want err, get nil")
-	}
-}
-
-func TestMergeScan(t *testing.T) {
-	c := getMergeTestConn(t)
-	defer c.Close()
-
-	testMergeKVScan(t, c)
-	testMergeHashKeyScan(t, c)
-	testMergeListKeyScan(t, c)
-	testMergeZSetKeyScan(t, c)
-	testMergeSetKeyScan(t, c)
-}
-
-func testMergeKVScan(t *testing.T, c *goredis.PoolConn) {
-	for i := 0; i < 20; i++ {
-		if _, err := c.Do("set", "default:testscan:"+fmt.Sprintf("%d", i), []byte("value")); err != nil {
-			t.Fatal(err)
-		}
-	}
-	checkMergeAdvanceScan(t, c, "KV")
-}
-
-func testMergeHashKeyScan(t *testing.T, c *goredis.PoolConn) {
-	for i := 0; i < 20; i++ {
-		if _, err := c.Do("hset", "default:testscan:"+fmt.Sprintf("%d", i), fmt.Sprintf("%d", i), []byte("value")); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	checkMergeAdvanceScan(t, c, "HASH")
-}
-
-func testMergeListKeyScan(t *testing.T, c *goredis.PoolConn) {
-	for i := 0; i < 20; i++ {
-		if _, err := c.Do("lpush", "default:testscan:"+fmt.Sprintf("%d", i), fmt.Sprintf("%d", i)); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	checkMergeAdvanceScan(t, c, "LIST")
-}
-
-func testMergeZSetKeyScan(t *testing.T, c *goredis.PoolConn) {
-	for i := 0; i < 20; i++ {
-		if _, err := c.Do("zadd", "default:testscan:"+fmt.Sprintf("%d", i), i, []byte("value")); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	checkMergeAdvanceScan(t, c, "ZSET")
-}
-
-func testMergeSetKeyScan(t *testing.T, c *goredis.PoolConn) {
-	for i := 0; i < 20; i++ {
-		if _, err := c.Do("sadd", "default:testscan:"+fmt.Sprintf("%d", i), fmt.Sprintf("%d", i)); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	checkMergeAdvanceScan(t, c, "SET")
 }
